pyproject.toml
[tool.poetry]
name = "coolprojectrealhype"
version = "0.1.0"
description = ""
authors = ["Mike <amaglichc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
SQLAlchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
bcrypt = "^4.1.3"
PyJWT = { extras = ["crypto"], version = "^2.8.0" }
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
